Updated the Property list a lot.

+++ server/src/customtags.cpp
@@ -323,18 +323,18 @@ const QString cVariant::toString() const
 	{
 		P_CHAR pChar = static_cast< P_CHAR >( d->value.ptr );
 		if( pChar )
-			return "0x" + QString::number( pChar->serial, 16 );
+			return "0x" + QString::number( (unsigned int)pChar->serial, 16 );
 		else
-			return "0x" + QString::number( INVALID_SERIAL, 16 );
+			return "0x" + QString::number( (unsigned int)INVALID_SERIAL, 16 );
 	}
 
 	if ( d->typ == Item )
 	{		
 		P_ITEM pItem = static_cast< P_ITEM >( d->value.ptr );
 		if( pItem )
-			return "0x" + QString::number( pItem->serial, 16 );
+			return "0x" + QString::number( (unsigned int)pItem->serial, 16 );
 		else
-			return "0x" + QString::number( INVALID_SERIAL, 16 );
+			return "0x" + QString::number( (unsigned int)INVALID_SERIAL, 16 );
 	}
 
 	if ( d->typ == Coord )
